# introduce struct BasicProperties for basicPublish