--- a/projects/RabbitMQ.Client/client/api/BasicGetResult.cs
+++ b/projects/RabbitMQ.Client/client/api/BasicGetResult.cs
@@ -53,7 +53,7 @@ namespace RabbitMQ.Client
         /// <param name="basicProperties">The Basic-class content header properties for the message.</param>
         /// <param name="body">The body</param>
         public BasicGetResult(ulong deliveryTag, bool redelivered, string exchange, string routingKey,
-            uint messageCount, IBasicProperties basicProperties, ReadOnlyMemory<byte> body)
+            uint messageCount, in ReadOnlyBasicProperties basicProperties, ReadOnlyMemory<byte> body)
         {
             DeliveryTag = deliveryTag;
             Redelivered = redelivered;
@@ -76,7 +76,7 @@ namespace RabbitMQ.Client
         /// <param name="body">The body</param>
         /// <param name="rentedArray">The rented array which body is part of.</param>
         public BasicGetResult(ulong deliveryTag, bool redelivered, string exchange, string routingKey,
-            uint messageCount, IBasicProperties basicProperties, ReadOnlyMemory<byte> body, byte[] rentedArray)
+            uint messageCount, in ReadOnlyBasicProperties basicProperties, ReadOnlyMemory<byte> body, byte[] rentedArray)
         {
             DeliveryTag = deliveryTag;
             Redelivered = redelivered;
@@ -91,7 +91,7 @@ namespace RabbitMQ.Client
         /// <summary>
         /// Retrieves the Basic-class content header properties for this message.
         /// </summary>
-        public IBasicProperties BasicProperties { get; }
+        public ReadOnlyBasicProperties BasicProperties { get; }
 
         /// <summary>
         /// Retrieves the body of this message.
@@ -130,7 +130,7 @@ namespace RabbitMQ.Client
         /// <inheritdoc />
         public void Dispose()
         {
-            if (!(_rentedArray is null))
+            if (_rentedArray is not null)
             {
                 ArrayPool<byte>.Shared.Return(_rentedArray);
             }