--- a/projects/RabbitMQ.Client/client/api/IBasicConsumer.cs
+++ b/projects/RabbitMQ.Client/client/api/IBasicConsumer.cs
@@ -94,7 +94,7 @@ namespace RabbitMQ.Client
             bool redelivered,
             string exchange,
             string routingKey,
-            IBasicProperties properties,
+            in ReadOnlyBasicProperties properties,
             ReadOnlyMemory<byte> body);
 
         /// <summary>