… projects/Benchmarks/ConsumerDispatching/AsyncBasicConsumerFake.cs
… projects/Benchmarks/ConsumerDispatching/ConsumerDispatcher.cs
… projects/Benchmarks/Networking/Networking_BasicDeliver_Commons.cs
… projects/Benchmarks/Networking/Networking_BasicDeliver_LongLivedConnection.cs
… projects/Benchmarks/WireFormatting/MethodFraming.cs
… projects/Benchmarks/WireFormatting/MethodSerialization.cs
… projects/RabbitMQ.Client/client/api/IBasicProperties.cs
… projects/RabbitMQ.Client/client/api/IStreamProperties.cs
… projects/RabbitMQ.Client/client/api/ReadonlyBasicProperties.cs
… projects/RabbitMQ.Client/client/events/BasicReturnEventArgs.cs
… projects/RabbitMQ.Client/client/events/EventingBasicConsumer.cs
… projects/RabbitMQ.Client/client/framing/BasicAck.cs
… projects/RabbitMQ.Client/client/framing/BasicCancel.cs
… projects/RabbitMQ.Client/client/framing/BasicConsume.cs
… projects/RabbitMQ.Client/client/framing/BasicGet.cs
… projects/RabbitMQ.Client/client/framing/BasicNack.cs
… projects/RabbitMQ.Client/client/framing/BasicProperties.cs
… projects/RabbitMQ.Client/client/framing/BasicPublish.cs
… projects/RabbitMQ.Client/client/framing/BasicQos.cs
… projects/RabbitMQ.Client/client/framing/BasicRecover.cs
… projects/RabbitMQ.Client/client/framing/BasicRecoverAsync.cs
… projects/RabbitMQ.Client/client/framing/BasicReject.cs
… projects/RabbitMQ.Client/client/framing/ChannelClose.cs
… projects/RabbitMQ.Client/client/framing/ChannelCloseOk.cs
… projects/RabbitMQ.Client/client/framing/ChannelFlowOk.cs
… projects/RabbitMQ.Client/client/framing/ChannelOpen.cs
… projects/RabbitMQ.Client/client/framing/ConfirmSelect.cs
… projects/RabbitMQ.Client/client/framing/ConnectionClose.cs
… projects/RabbitMQ.Client/client/framing/ConnectionCloseOk.cs
… projects/RabbitMQ.Client/client/framing/ConnectionOpen.cs