## Proposed Changes

Introduces a struct BasicProperties and opens up the header implementation to a public.

Changes in detail:
- Replaced IBasicProperties in the receiving area with ReadOnlyBasicProperties
- Replaced BasicProperties class with a struct version
- Modified BasicPublish to be generic (TProperties : IBasicProperties, IAmqpHeader)
- Removed CreateBasicProperties method, replacement is 'new BasicProperties()'
- Introduce public IAmqpHeader interface. (Can be implemented by anyone and used as header properties)
- Removed IStreamProperties, IContentHeader (Not needed for anything)
- Aligned the name for getting the buffer size: GetRequiredPayloadBufferSize -> GetRequiredBufferSize
- Aligned the name for writing to the buffer: WritePropertiesTo -> WriteTo

## Types of Changes

- [ ] Bug fix (non-breaking change which fixes issue #NNNN)
- [x] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause an observable behavior change in existing systems)
- [ ] Documentation improvements (corrections, new content, etc)
- [ ] Cosmetic change (whitespace, formatting, etc)

## Checklist

- [x] I have read the `CONTRIBUTING.md` document
- [x] I have signed the CA (see https://cla.pivotal.io/sign/rabbitmq)
- [x] All tests pass locally with my changes
- [x] I have added tests that prove my fix is effective or that my feature works
- [ ] I have added necessary documentation (if appropriate)
- [x] Any dependent changes have been merged and published in related repositories

## Further Comments

I'm going to post a few benchmark numbers as soon as I get around to run them. (soon)