@michaelklishin 2021-10-08
Thank you. Looks like this is still a WIP? I run into compilation errors:

```
/Users/antares/Development/RabbitMQ/dotnet_client.git/projects/Unit/TestConnectionRecovery.cs(783,20): error CS0411: The type arguments for method 'IModel.BasicPublish<TProperties>(string, string, in TProperties, ReadOnlyMemory<byte>, bool)' cannot be inferred from the usage. Try specifying the type arguments explicitly. [/path/to/dotnet_client.git/projects/Unit/Unit.csproj]
/path/to/dotnet_client.git/projects/Unit/TestConnectionRecovery.cs(798,37): error CS1061: 'IModel' does not contain a definition for 'CreateBasicProperties' and no accessible extension method 'CreateBasicProperties' accepting a first argument of type 'IModel' could be found (are you missing a using directive or an assembly reference?) [/path/to/dotnet_client.git/projects/Unit/Unit.csproj]
```