@bollhals 2021-10-08
Ah probably my last rebase against main... sorry for that. I'll update it when I ran the benchmarks👍🏼