@bollhals 2021-10-16
|                    Method | When   |            Runtime |      Mean |    Error |   StdDev | Code Size |
|-------------------------- |--------|------------------- |----------:|---------:|---------:|----------:|
| BasicPublishWriteNonEmpty | Before |      .NET Core 3.1 | 324.15 ns | 2.803 ns | 2.622 ns |   5,638 B |
| BasicPublishWriteNonEmpty | After  |      .NET Core 3.1 | 327.46 ns | 2.391 ns | 2.236 ns |   5,697 B |
|         BasicPublishWrite | Before |      .NET Core 3.1 | 118.49 ns | 1.718 ns | 1.607 ns |   5,638 B |
|         BasicPublishWrite | After  |      .NET Core 3.1 |  86.68 ns | 0.507 ns | 0.450 ns |   2,651 B |
|   BasicPublishMemoryWrite | Before |      .NET Core 3.1 |  72.80 ns | 0.386 ns | 0.342 ns |   6,127 B |
|   BasicPublishMemoryWrite | After  |      .NET Core 3.1 |  44.78 ns | 0.288 ns | 0.270 ns |   2,594 B |
| BasicPublishWriteNonEmpty | Before | .NET Framework 4.8 | 438.34 ns | 3.226 ns | 3.018 ns |   6,831 B |
| BasicPublishWriteNonEmpty | After  | .NET Framework 4.8 | 443.86 ns | 4.187 ns | 3.712 ns |   6,849 B |
|         BasicPublishWrite | Before | .NET Framework 4.8 | 220.12 ns | 0.528 ns | 0.468 ns |   6,703 B |
|         BasicPublishWrite | After  | .NET Framework 4.8 | 182.63 ns | 0.969 ns | 0.906 ns |   3,006 B |
|   BasicPublishMemoryWrite | Before | .NET Framework 4.8 | 153.67 ns | 1.525 ns | 1.427 ns |   7,044 B |
|   BasicPublishMemoryWrite | After  | .NET Framework 4.8 | 119.55 ns | 0.652 ns | 0.610 ns |   3,347 B |

|              Method | When |     Mean |   Error |  StdDev | Ratio |    Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------- |------|---------:|--------:|--------:|------:|---------:|------:|------:|----------:|
| Publish_Hello_World | Before |149.2 ms | 2.78 ms | 2.60 ms |  1.00 | 750.0000 |     - |     - |   4.33 MB |
| Publish_Hello_World | After | 148.1 ms | 2.94 ms | 4.66 ms |  1.00 | 750.0000 |     - |     - |   4.02 MB |

Before
![Before](https://user-images.githubusercontent.com/2437596/137602678-9365e364-2659-4cc6-a97d-6e2f6e1618f3.jpg)

After
![after](https://user-images.githubusercontent.com/2437596/137602685-b0c0737b-eee1-4632-a76b-f7c4fa32cdf1.jpg)


Summary:
Pure serialization performance is slight worse, but overall the performance is better and there is less memory allocations.
Also performance without properties is significant faster.

Remaining memory allocation for sending and receiving is mostly string creation.