@michaelklishin 2021-12-27
I had to update `RabbitMQ.Client.Unit.APIApproval` expectations, running the rest of the suites.