@bollhals 2021-12-28
Whoops, my idea was just to review, as this change does have impact on some basic API that will require some turtorials/docu update. 

Were you aware of this?